forge fmt

+++ contracts/src/v0.8/llo-feeds/v0.5.0/test/verifier/BaseVerifierTest.t.sol
@@ -1,19 +1,22 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.19;
 
-import {Test} from "forge-std/Test.sol";
-import {VerifierProxy} from "../../VerifierProxy.sol";
-import {IERC165} from "@openzeppelin/contracts@4.8.3/interfaces/IERC165.sol";
-import {IVerifier} from "../../interfaces/IVerifier.sol";
-import {ErroredVerifier} from "../mocks/ErroredVerifier.sol";
-import {Verifier} from "../../Verifier.sol";
-import {Strings} from "@openzeppelin/contracts@4.9.6/utils/Strings.sol";
 import {AccessControllerInterface} from "../../../../shared/interfaces/AccessControllerInterface.sol";
-import {FeeManager} from "../../FeeManager.sol";
-import {Common} from "../../../libraries/Common.sol";
+
 import {ERC20Mock} from "../../../../shared/mocks/ERC20Mock.sol";
 import {WERC20Mock} from "../../../../shared/mocks/WERC20Mock.sol";
+import {Common} from "../../../libraries/Common.sol";
+import {FeeManager} from "../../FeeManager.sol";
+
 import {RewardManager} from "../../RewardManager.sol";
+import {Verifier} from "../../Verifier.sol";
+import {VerifierProxy} from "../../VerifierProxy.sol";
+import {IVerifier} from "../../interfaces/IVerifier.sol";
+import {ErroredVerifier} from "../mocks/ErroredVerifier.sol";
+import {IERC165} from "@openzeppelin/contracts@4.8.3/interfaces/IERC165.sol";
+
+import {Strings} from "@openzeppelin/contracts@4.9.6/utils/Strings.sol";
+import {Test} from "forge-std/Test.sol";
 
 contract BaseTest is Test {
   uint256 internal constant MAX_ORACLES = 31;
@@ -114,7 +117,9 @@ contract BaseTest is Test {
     }
   }
 
-  function _getSigners(uint256 numSigners) internal view returns (Signer[] memory) {
+  function _getSigners(
+    uint256 numSigners
+  ) internal view returns (Signer[] memory) {
     Signer[] memory signers = new Signer[](numSigners);
     for (uint256 i; i < numSigners; i++) {
       signers[i] = s_signers[i];
@@ -122,7 +127,9 @@ contract BaseTest is Test {
     return signers;
   }
 
-  function _getSignerAddresses(Signer[] memory signers) internal view returns (address[] memory) {
+  function _getSignerAddresses(
+    Signer[] memory signers
+  ) internal view returns (address[] memory) {
     address[] memory signerAddrs = new address[](signers.length);
     for (uint256 i = 0; i < signerAddrs.length; i++) {
       signerAddrs[i] = s_signers[i].signerAddress;
@@ -150,19 +157,20 @@ contract BaseTest is Test {
     return (rs, ss, bytes32(vs));
   }
 
-  function _encodeReport(V1Report memory report) internal pure returns (bytes memory) {
-    return
-      abi.encode(
-        report.feedId,
-        report.observationsTimestamp,
-        report.median,
-        report.bid,
-        report.ask,
-        report.blocknumberUpperBound,
-        report.upperBlockhash,
-        report.blocknumberLowerBound,
-        report.currentBlockTimestamp
-      );
+  function _encodeReport(
+    V1Report memory report
+  ) internal pure returns (bytes memory) {
+    return abi.encode(
+      report.feedId,
+      report.observationsTimestamp,
+      report.median,
+      report.bid,
+      report.ask,
+      report.blocknumberUpperBound,
+      report.upperBlockhash,
+      report.blocknumberLowerBound,
+      report.currentBlockTimestamp
+    );
   }
 
   function _generateV1EncodedBlob(
@@ -171,11 +179,8 @@ contract BaseTest is Test {
     Signer[] memory signers
   ) internal pure returns (bytes memory) {
     bytes memory reportBytes = _encodeReport(report);
-    (bytes32[] memory rs, bytes32[] memory ss, bytes32 rawVs) = _generateSignerSignatures(
-      reportBytes,
-      reportContext,
-      signers
-    );
+    (bytes32[] memory rs, bytes32[] memory ss, bytes32 rawVs) =
+      _generateSignerSignatures(reportBytes, reportContext, signers);
     return abi.encode(reportContext, reportBytes, rs, ss, rawVs);
   }
 
@@ -193,7 +198,7 @@ contract BaseTest is Test {
   ) internal pure returns (bytes32) {
     // Convert addresses to bytes array to match configurator
     bytes[] memory signersAsBytes = new bytes[](signers.length);
-    for (uint i; i < signers.length; ++i) {
+    for (uint256 i; i < signers.length; ++i) {
       signersAsBytes[i] = abi.encodePacked(signers[i]);
     }
 
@@ -229,31 +234,29 @@ contract BaseTest is Test {
     uint64 blocknumberLowerBound,
     uint32 currentBlockTimestamp
   ) internal pure returns (V1Report memory) {
-    return
-      V1Report({
-        feedId: feedId,
-        observationsTimestamp: observationsTimestamp,
-        median: median,
-        bid: bid,
-        ask: ask,
-        blocknumberUpperBound: blocknumberUpperBound,
-        upperBlockhash: upperBlockhash,
-        blocknumberLowerBound: blocknumberLowerBound,
-        currentBlockTimestamp: currentBlockTimestamp
-      });
+    return V1Report({
+      feedId: feedId,
+      observationsTimestamp: observationsTimestamp,
+      median: median,
+      bid: bid,
+      ask: ask,
+      blocknumberUpperBound: blocknumberUpperBound,
+      upperBlockhash: upperBlockhash,
+      blocknumberLowerBound: blocknumberLowerBound,
+      currentBlockTimestamp: currentBlockTimestamp
+    });
   }
 
   function _ccipReadURL(bytes32 feedId, uint256 commitmentBlock) internal pure returns (string memory url) {
-    return
-      string(
-        abi.encodePacked(
-          SERVER_URL,
-          "?feedIDHex=",
-          Strings.toHexString(uint256(feedId)),
-          "&L2Blocknumber=",
-          Strings.toString(commitmentBlock)
-        )
-      );
+    return string(
+      abi.encodePacked(
+        SERVER_URL,
+        "?feedIDHex=",
+        Strings.toHexString(uint256(feedId)),
+        "&L2Blocknumber=",
+        Strings.toString(commitmentBlock)
+      )
+    );
   }
 }
 
@@ -310,10 +313,7 @@ contract BaseTestWithConfiguredVerifierAndFeeManager is BaseTest {
     );
 
     s_verifier.setConfig(
-      v1ConfigDigest,
-      _getSignerAddresses(signers),
-      FAULT_TOLERANCE,
-      new Common.AddressAndWeight[](0)
+      v1ConfigDigest, _getSignerAddresses(signers), FAULT_TOLERANCE, new Common.AddressAndWeight[](0)
     );
 
     v3ConfigDigest = _configDigestFromConfigData(
@@ -330,10 +330,7 @@ contract BaseTestWithConfiguredVerifierAndFeeManager is BaseTest {
     );
 
     s_verifier.setConfig(
-      v3ConfigDigest,
-      _getSignerAddresses(signers),
-      FAULT_TOLERANCE,
-      new Common.AddressAndWeight[](0)
+      v3ConfigDigest, _getSignerAddresses(signers), FAULT_TOLERANCE, new Common.AddressAndWeight[](0)
     );
 
     link = new ERC20Mock(18);
@@ -346,19 +343,20 @@ contract BaseTestWithConfiguredVerifierAndFeeManager is BaseTest {
     rewardManager.setFeeManager(address(feeManager));
   }
 
-  function _encodeReport(V3Report memory report) internal pure returns (bytes memory) {
-    return
-      abi.encode(
-        report.feedId,
-        report.observationsTimestamp,
-        report.validFromTimestamp,
-        report.nativeFee,
-        report.linkFee,
-        report.expiresAt,
-        report.benchmarkPrice,
-        report.bid,
-        report.ask
-      );
+  function _encodeReport(
+    V3Report memory report
+  ) internal pure returns (bytes memory) {
+    return abi.encode(
+      report.feedId,
+      report.observationsTimestamp,
+      report.validFromTimestamp,
+      report.nativeFee,
+      report.linkFee,
+      report.expiresAt,
+      report.benchmarkPrice,
+      report.bid,
+      report.ask
+    );
   }
 
   function _generateV3EncodedBlob(
@@ -367,45 +365,42 @@ contract BaseTestWithConfiguredVerifierAndFeeManager is BaseTest {
     Signer[] memory signers
   ) internal pure returns (bytes memory) {
     bytes memory reportBytes = _encodeReport(report);
-    (bytes32[] memory rs, bytes32[] memory ss, bytes32 rawVs) = _generateSignerSignatures(
-      reportBytes,
-      reportContext,
-      signers
-    );
+    (bytes32[] memory rs, bytes32[] memory ss, bytes32 rawVs) =
+      _generateSignerSignatures(reportBytes, reportContext, signers);
     return abi.encode(reportContext, reportBytes, rs, ss, rawVs);
   }
 
   function _generateV1Report() internal view returns (V1Report memory) {
-    return
-      _createV1Report(
-        FEED_ID,
-        OBSERVATIONS_TIMESTAMP,
-        MEDIAN,
-        BID,
-        ASK,
-        BLOCKNUMBER_UPPER_BOUND,
-        bytes32(blockhash(BLOCKNUMBER_UPPER_BOUND)),
-        BLOCKNUMBER_LOWER_BOUND,
-        uint32(block.timestamp)
-      );
+    return _createV1Report(
+      FEED_ID,
+      OBSERVATIONS_TIMESTAMP,
+      MEDIAN,
+      BID,
+      ASK,
+      BLOCKNUMBER_UPPER_BOUND,
+      bytes32(blockhash(BLOCKNUMBER_UPPER_BOUND)),
+      BLOCKNUMBER_LOWER_BOUND,
+      uint32(block.timestamp)
+    );
   }
 
   function _generateV3Report() internal view returns (V3Report memory) {
-    return
-      V3Report({
-        feedId: FEED_ID_V3,
-        observationsTimestamp: OBSERVATIONS_TIMESTAMP,
-        validFromTimestamp: uint32(block.timestamp),
-        nativeFee: uint192(DEFAULT_REPORT_NATIVE_FEE),
-        linkFee: uint192(DEFAULT_REPORT_LINK_FEE),
-        expiresAt: uint32(block.timestamp),
-        benchmarkPrice: MEDIAN,
-        bid: BID,
-        ask: ASK
-      });
+    return V3Report({
+      feedId: FEED_ID_V3,
+      observationsTimestamp: OBSERVATIONS_TIMESTAMP,
+      validFromTimestamp: uint32(block.timestamp),
+      nativeFee: uint192(DEFAULT_REPORT_NATIVE_FEE),
+      linkFee: uint192(DEFAULT_REPORT_LINK_FEE),
+      expiresAt: uint32(block.timestamp),
+      benchmarkPrice: MEDIAN,
+      bid: BID,
+      ask: ASK
+    });
   }
 
-  function _generateReportContext(bytes32 configDigest) internal pure returns (bytes32[3] memory) {
+  function _generateReportContext(
+    bytes32 configDigest
+  ) internal pure returns (bytes32[3] memory) {
     bytes32[3] memory reportContext;
     reportContext[0] = configDigest;
     reportContext[1] = bytes32(abi.encode(uint32(5), uint8(1)));
@@ -437,12 +432,7 @@ contract BaseTestWithConfiguredVerifierAndFeeManager is BaseTest {
     changePrank(originalAddr);
   }
 
-  function _verifyBulk(
-    bytes[] memory payload,
-    address feeAddress,
-    uint256 wrappedNativeValue,
-    address sender
-  ) internal {
+  function _verifyBulk(bytes[] memory payload, address feeAddress, uint256 wrappedNativeValue, address sender) internal {
     address originalAddr = msg.sender;
     changePrank(sender);
 
@@ -485,10 +475,7 @@ contract BaseTestWithMultipleConfiguredDigests is BaseTestWithConfiguredVerifier
       bytes("")
     );
     s_verifier.setConfig(
-      s_configDigestTwo,
-      _getSignerAddresses(secondSetOfSigners),
-      FAULT_TOLERANCE_TWO,
-      new Common.AddressAndWeight[](0)
+      s_configDigestTwo, _getSignerAddresses(secondSetOfSigners), FAULT_TOLERANCE_TWO, new Common.AddressAndWeight[](0)
     );
 
     // Verifier 1, Feed 1, Config 3
@@ -526,10 +513,7 @@ contract BaseTestWithMultipleConfiguredDigests is BaseTestWithConfiguredVerifier
       bytes("")
     );
     s_verifier.setConfig(
-      s_configDigestFour,
-      _getSignerAddresses(signers),
-      FAULT_TOLERANCE,
-      new Common.AddressAndWeight[](0)
+      s_configDigestFour, _getSignerAddresses(signers), FAULT_TOLERANCE, new Common.AddressAndWeight[](0)
     );
 
     // Verifier 2, Feed 3, Config 1
@@ -547,10 +531,7 @@ contract BaseTestWithMultipleConfiguredDigests is BaseTestWithConfiguredVerifier
       bytes("")
     );
     s_verifier_2.setConfig(
-      s_configDigestFive,
-      _getSignerAddresses(signers),
-      FAULT_TOLERANCE,
-      new Common.AddressAndWeight[](0)
+      s_configDigestFive, _getSignerAddresses(signers), FAULT_TOLERANCE, new Common.AddressAndWeight[](0)
     );
   }
 }
